termpature unit setting added fahrenheit interval not working

+++ lib/widgets/custom_modal_sheet/custom_vertical_weight_slider.dart
@@ -84,7 +84,7 @@ class _SettingValueSliderState extends State<SettingValueSlider> {
         );
       case ParameterType.coffeeAmount:
         return WeightSliderController(
-          initialWeight: _settingsSliderProvider.tempCoffeeAmount!.toDouble(),
+          initialWeight: _settingsSliderProvider.tempCoffeeAmount!,
           interval: 0.1,
         );
       case ParameterType.waterAmount:
